+++ src/registrationSystem/models/__init__.py
@@ -1,0 +1,12 @@
+# flake8: noqa
+
+from .river_rafting import (
+    RiverraftingUser,
+    RiverraftingTeam,
+    RiverRaftingRaffleState,
+    InterestCheck
+)
+
+from .common import (
+    CommonUserManager
+)
