+++ src/registrationSystem/models/common/__init__.py
@@ -1,0 +1,4 @@
+from .abstract_user import AbstractBaseUser
+from .abstract_group import AbstractGroup
+from .abstract_singleton_model import AbstractSingletonModel
+from .common_user_manager import CommonUserManager
