+++ src/registrationSystem/models/common/abstract_singleton_model.py
@@ -1,0 +1,31 @@
+
+class AbstractSingletonModel(models.Model):
+    """
+    Taken from
+    https://stackoverflow.com/questions/49735906/how-to-implement-singleton-in-django
+    """
+    class Meta:
+        abstract = True
+
+    def save(self, *args, **kwargs):
+        """
+        Save object to the database. Removes all other entries if there
+        are any.
+        """
+        self.__class__.objects.exclude(id=self.id).delete()
+        super(AbstractSingletonModel, self).save(*args, **kwargs)
+
+    def delete(self, *args, **kwargs):
+        pass
+    
+    @classmethod
+    def load(cls):
+        """
+        Load object from the database. Failing that, create a new empty
+        (default) instance of the object and return it (without saving it
+        to the database).
+        """
+        try:
+            return cls.objects.get()
+        except cls.DoesNotExist:
+            return cls()
